float perfect

+++ src/util.odin
@@ -2,7 +2,7 @@ package main
 
 import "core:math"
 
-PIf: f32 = 3.14159265358979323846
+PIf: f32 = f32(3.14159265358979323846)
 PId: f64 = 3.14159265358979323846264338327950288
 
 SIN_TABLE: [65536]f32
@@ -18,7 +18,7 @@ sinr :: proc(rad: f32) -> f32 {
 }
 
 sin :: proc(deg: f32) -> f32 {
-    rad := deg * PIf / 180.0
+    rad := deg * PIf / f32(180.0)
     return SIN_TABLE[int(rad * 10430.378) & 65535]
 }
 
@@ -27,6 +27,6 @@ cosr :: proc(rad: f32) -> f32 {
 }
 
 cos :: proc(deg: f32) -> f32 {
-    rad := deg * PIf / 180.0
+    rad := deg * PIf / f32(180.0)
     return SIN_TABLE[int(rad * 10430.378 + 16384.0) & 65535]
 }
